+++ app/src/main/java/com/aylanetworks/agilelink/device/AylaEVBDevice.java
@@ -86,7 +86,7 @@ public class AylaEVBDevice extends GenericDevice implements View.OnClickListener
 
     public boolean isGreenLEDOn() {
         AylaProperty prop = _device.getProperty(PROPERTY_GREEN_LED);
-        if (prop != null && prop.getValue() != null && prop.getValue() != 0) {
+        if (prop != null && prop.getValue() != null && (int) prop.getValue() != 0) {
             return true;
         }
         return false;
@@ -94,7 +94,7 @@ public class AylaEVBDevice extends GenericDevice implements View.OnClickListener
 
     public boolean isBlueLEDOn() {
         AylaProperty prop = _device.getProperty(PROPERTY_BLUE_LED);
-        if (prop != null && prop.getValue() != null && prop.getValue() != 0) {
+        if (prop != null && prop.getValue() != null && (int) prop.getValue() != 0) {
             return true;
         }
         return false;
